Separate cli_parser.py into two modules
In order to avoid import cycles for future cli changes as well as to improve
the high coupling of the cli_parser module (which was over 2k lines),
create a new module which will hold the parser methods and another module
to contain the config/definitions of the Args, Commands, Actions, etc
